Refactoring to better merge with master
