+++ server/model/model.go
@@ -63,7 +63,7 @@ type CommentSource struct {
 
 // Comment denotes a comment by a user in the system.
 type Comment struct {
-	ID           bson.ObjectId          `json:"id" bson:"_id"`
+	ID           bson.ObjectId          `json:"_id" bson:"_id"`
 	UserID       bson.ObjectId          `json:"user_id" bson:"user_id"`
 	AssetID      bson.ObjectId          `json:"asset_id" bson:"asset_id"`
 	ParentID     bson.ObjectId          `json:"parent_id,omitempty" bson:"parent_d,omitempty"`
